feat(deps): upgrade to tree-sitter 0.27
tree-sitter 0.27 requires tree-sitter-language ^0.1.8, and 0.1.8 relocated the 0.26 wasm
shims under wasm/unsupported/ so they no longer compile against it; the two pins move
together or neither does. 0.27 replaces the 0.26 wasm integration's non-thread-safe
allocator and incomplete libc with its own wasm libc, forwarding C allocation to Rust's
global allocator.

Node::child_count now returns u32 upstream, so the wrapper casts to its declared usize.
That is the only source change the runtime bump requires.

+++ crates/ts-pack-core/src/parsing.rs
@@ -375,7 +375,7 @@ impl Node {
     /// Total number of children (including unnamed).
     #[must_use]
     pub fn child_count(&self) -> usize {
-        self.raw.child_count()
+        self.raw.child_count() as usize
     }
 
     /// Return the i-th named child of this node, if any.
